apps/nest-neo4j-api/src/app/users/users.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { UsersService } from './users.service';
import { Neo4jService } from '../neo4j/neo4j.service';
import { User } from './entities/user.entity';
describe('UsersService', () => {
  let service: UsersService;
  let neo4jService: Neo4jService;

  beforeEach(async () => {
    const module: TestingModule = await Test.createTestingModule({
      providers: [
        UsersService,
        {
          provide: Neo4jService,
          useValue: {
            // Mock the methods of Neo4jService that are used in UsersService
            // For example:
            // createUser: jest.fn(),
            // findUserByUsername: jest.fn(),
            // findAllUsers: jest.fn(),
            // updateUser: jest.fn(),
            // removeUser: jest.fn(),
          },
        },
      ],
    }).compile();

    service = module.get<UsersService>(UsersService);
    neo4jService = module.get<Neo4jService>(Neo4jService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });

  describe('create', () => {
    it('should create a new user', async () => {
      // Mock the necessary dependencies and test the create method
      // For example:
      // const createUserDto: CreateUserDto = { username: 'test', email: 'test@example.com', password: 'password' };
      // const createdUser: User = { id: '1', ...createUserDto };
      // jest.spyOn(neo4jService, 'createUser').mockResolvedValue(createdUser);
      // const result = await service.create(createUserDto.username, createUserDto.email, createUserDto.password);
      // expect(result).toEqual(createdUser);
    });
  });

  describe('findOne', () => {
    it('should find a user by username', async () => {
      // Mock the necessary dependencies and test the findOne method
      // For example:
      // const username = 'test';
      // const foundUser: User = { id: '1', username, email: 'test@example.com', password: 'password' };
      // jest.spyOn(neo4jService, 'findUserByUsername').mockResolvedValue(foundUser);
      // const result = await service.findOne(username);
      // expect(result).toEqual(foundUser);
    });
  });

  describe('findAll', () => {
    it('should find all users based on properties', async () => {
      // Mock the necessary dependencies and test the findAll method
      // For example:
      // const properties = { email: 'test@example.com' };
      // const foundUsers: User[] = [{ id: '1', username: 'test', ...properties }];
      // jest.spyOn(neo4jService, 'findAllUsers').mockResolvedValue(foundUsers);
      // const result = await service.findAll(properties);
      // expect(result).toEqual(foundUsers);
    });
  });

  describe('update', () => {
    it('should update a user', async () => {
      // Mock the necessary dependencies and test the update method
      // For example:
      // const id = '1';
      // const updateUserDto: Partial<CreateUserDto> = { email: 'updated@example.com' };
      // const updatedUser: User = { id, username: 'test', email: updateUserDto.email, password: 'password' };
      // jest.spyOn(neo4jService, 'updateUser').mockResolvedValue(updatedUser);
      // const result = await service.update(id, updateUserDto);
      // expect(result).toEqual(updatedUser);
    });
  });

  describe('remove', () => {
    it('should remove a user', async () => {
      // Mock the necessary dependencies and test the remove method
      // For example:
      // const id = '1';
      // jest.spyOn(neo4jService, 'removeUser').mockResolvedValue(undefined);
      // await expect(service.remove(id)).resolves.toBeUndefined();
    });
  });
});
